// src/app/dto/model/offreImmobilier.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OffreImmobilier { 
    id?: number;
    typeBien: OffreImmobilier.TypeBienEnum;
    adresse: string;
    prix: number;
    description?: string;
    listFavoris?: number;
    imageUrls?: Array<string>;
    statutPropriete?: OffreImmobilier.StatutProprieteEnum;
    ameublement?: OffreImmobilier.AmeublementEnum;
    chambres?: OffreImmobilier.ChambresEnum;
    sallesDeBain?: OffreImmobilier.SallesDeBainEnum;
    surface?: number;
    agePropriete?: OffreImmobilier.AgeProprieteEnum;
}
export namespace OffreImmobilier {
    export type TypeBienEnum = 'MAISON' | 'APPARTEMENT' | 'VILLA' | 'LOFT';
    export const TypeBienEnum = {
        Maison: 'MAISON' as TypeBienEnum,
        Appartement: 'APPARTEMENT' as TypeBienEnum,
        Villa: 'VILLA' as TypeBienEnum,
        Loft: 'LOFT' as TypeBienEnum
    };
    export type StatutProprieteEnum = 'DISPONIBLE' | 'NON_DISPONIBLE' | 'RESERVE';
    export const StatutProprieteEnum = {
        Disponible: 'DISPONIBLE' as StatutProprieteEnum,
        NonDisponible: 'NON_DISPONIBLE' as StatutProprieteEnum,
        Reserve: 'RESERVE' as StatutProprieteEnum
    };
    export type AmeublementEnum = 'SEMI_EQUIPE' | 'NON_EQUIPE' | 'EQUIPE';
    export const AmeublementEnum = {
        SemiEquipe: 'SEMI_EQUIPE' as AmeublementEnum,
        NonEquipe: 'NON_EQUIPE' as AmeublementEnum,
        Equipe: 'EQUIPE' as AmeublementEnum
    };
    export type ChambresEnum = 'S1' | 'S2' | 'S3' | 'S4' | 'S5' | 'S6';
    export const ChambresEnum = {
        S1: 'S1' as ChambresEnum,
        S2: 'S2' as ChambresEnum,
        S3: 'S3' as ChambresEnum,
        S4: 'S4' as ChambresEnum,
        S5: 'S5' as ChambresEnum,
        S6: 'S6' as ChambresEnum
    };
    export type SallesDeBainEnum = 'B1' | 'B2' | 'B3' | 'B4' | 'B5' | 'B6';
    export const SallesDeBainEnum = {
        B1: 'B1' as SallesDeBainEnum,
        B2: 'B2' as SallesDeBainEnum,
        B3: 'B3' as SallesDeBainEnum,
        B4: 'B4' as SallesDeBainEnum,
        B5: 'B5' as SallesDeBainEnum,
        B6: 'B6' as SallesDeBainEnum
    };
    export type AgeProprieteEnum = 'MOINS_DE_1_AN' | 'MOINS_DE_3_ANS' | 'MOINS_DE_10_ANS' | 'PLUS_DE_10_ANS';
    export const AgeProprieteEnum = {
        MoinsDe1An: 'MOINS_DE_1_AN' as AgeProprieteEnum,
        MoinsDe3Ans: 'MOINS_DE_3_ANS' as AgeProprieteEnum,
        MoinsDe10Ans: 'MOINS_DE_10_ANS' as AgeProprieteEnum,
        PlusDe10Ans: 'PLUS_DE_10_ANS' as AgeProprieteEnum
    };
}